requirements.txt
# API
fastapi==0.115.5
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1

# Dashboard
streamlit==1.41.1
requests==2.32.3
pandas==2.2.3
plotly==5.24.1

# Pipeline
prefect==3.1.6